fix: moved toSentenceCase in its own file
