+++ src/components/Example.tsx
@@ -12,7 +12,7 @@ import {
 import Link from "next/link";
 import { ExternalLinkIcon } from "@chakra-ui/icons";
 
-import { toSentenceCase } from "@utils";
+import { toSentenceCase } from "@utils/toSentenceCase";
 import { ResizableFrame } from "@components/ResizableFrame";
 import { CodeSample } from "@components/CodeSample";
